+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Versions
 
+## 1.7.3
+
+- Show L-BTC balance changes when sending backup by telegram
+
 ## 1.7.2
 
 - Ignore inline comments in peerswap.conf